msg/Message: use an unsigned type for get_data_len()
to avoid warnings like:

../src/osd/PrimaryLogPG.cc:2123:27: warning: comparison of integer expressions of different signedness: ‘off_t’ {aka ‘long int’} and ‘uint64_t’ {aka ‘long unsigned int’} [-Wsign-compare]
 2123 |         m->get_data_len() > cct->_conf->osd_max_write_size << 20) {
      |         ~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

after change the size options' type to unsigned or uint64_t.

+++ src/msg/Message.h
@@ -430,7 +430,7 @@ public:
       byte_throttler->put(data.length());
     bl = std::move(data);
   }
-  off_t get_data_len() const { return data.length(); }
+  uint32_t get_data_len() const { return data.length(); }
 
   void set_recv_stamp(utime_t t) { recv_stamp = t; }
   const utime_t& get_recv_stamp() const { return recv_stamp; }
